done ex5-15

+++ exercises/006-conversion/conversion.go
@@ -3,8 +3,6 @@
 // https://go.dev/tour/basics/13
 // The expression T(v) converts the value v to the type T.
 
-// I AM STILL GOING
-
 package main
 
 import "fmt"
@@ -17,5 +15,5 @@ func main() {
 	// between variables they must be of the same type! Convert one of
 	// the variables in the Println so that the result of 1/3+2+3 is printed out.
 	// Don't add lines or change the lines above!
-	fmt.Println(oneThird + twoPlusThree)
+	fmt.Println(oneThird + float64(twoPlusThree))
 }
